CardLongAnswerParameterInput y CardNumberParameterInputs modularizados y pasados a material
