+++ components/FormBuilder/CardDefaultParameterInputs.js
@@ -1,0 +1,6 @@
+import { createElement } from "react";
+
+// specify the inputs required for a string type object
+export default function CardDefaultParameterInputs({ parameters, onChange }) {
+  return createElement("div", null);
+}
